server/requirements.txt
psycopg2-binary
databases
databases[postgresql]
passlib
python-jose
aiohttp